% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMR.R
\name{PCMR}
\alias{PCMR}
\title{Pleiotropy Clustering Mendelian Randomiztion(PCMR)}
\usage{
PCMR(
  beta_ex,
  beta_ex_se,
  beta_out,
  beta_out_se,
  num_gamma,
  model = "2",
  isIntact = FALSE,
  num_b = 2,
  rho = 0,
  gamma = NA,
  pi_gamma = NA,
  b = NA,
  pi_b = NA,
  sigma2 = NA,
  gSigma2 = NA,
  paraUpdate = NA,
  thred = 1e-06,
  maxIteration = 1000
)
}
\arguments{
\item{beta_ex}{A numeric vector of beta-coefficient values for genetic associations with the exposure.}

\item{beta_ex_se}{The standard errors associated with the beta-coefficients beta_ex.}

\item{beta_out}{A numeric vector of beta-coefficient values for genetic associations with the outcome.}

\item{beta_out_se}{The standard errors associated with the beta-coefficients beta_out.}

\item{num_gamma}{The number of classes of correlated pleiotropy and causal effect. 1 represents assuming IVs only have causal effect. 2 represents assuming IVs have correlated pleiotropy besides causal effect.}

\item{model}{A character representing different model. "1" represents PCMR model with random sigma2 and gSigma2. "2" represents PCMR model with random sigma2 and fixed gSigma2."3" represents PCMR model with fixed sigma2 and random gSigma2."4" represents PCMR model with fixed sigma2 and gSigma2. Default is "2".}

\item{isIntact}{Whether using the intact model to fitting. Default is FALSE.}

\item{num_b}{The number of classes of uncorrelated pleiotropy.}

\item{rho}{The covariance of beta_ex and beta_out. Default is zero.}

\item{gamma}{The initial value of correlated pleiotropy and causal effect.  A numeric vector of same length as num_gamma. The default is quantile.}

\item{pi_gamma}{The initial proportion of gamma. The default is the same ratio for each gamma.}

\item{b}{A numeric value of uncorrelated pleiotropy. The default is zero.}

\item{pi_b}{The initial proportion of b.The default is the same ratio for each sigma2.}

\item{sigma2}{A bivector and the initial variance of b. The default is bivector of zero.}

\item{gSigma2}{The intitial variance of gamma. The default is zero.}

\item{paraUpdate}{A list containing which variables can be updated.}

\item{thred}{The thredhold of convergence contidion. The default is 1e-6.}

\item{maxIteration}{The max iteration. The default is 1000.}
}
\value{
gamma The estimated of correlated pleiotropy and causal effect.

gSigma2 The estimated variance of gamma.

pi_gamma The estimated proportion of gamma.

b The estimated of uncorrelated pleiotropy.

sigma2 The estimated variance of b.

pi_b The estimated proportion of sigma2.
}
\description{
Pleiotropy Clustering Mendelian Randomiztion(PCMR)
}
\references{
\url{urlrul}
}
\author{
bintang<tangbin0513@gmail.com>
}
